Don't add a default kernel cmdline
Specifying a console device makes sense when you know what device is
going to be used and don't want the kernel to pick the first
reasonable device found. stmgr should not assume that it knows what
the situation looks like on the target system.

Related to https://git.glasklar.is/system-transparency/core/system-transparency/-/issues/200