chore: Upgrade and adapt for go-diskfs v1.6.0
The go-diskfs git-commit 5e0c7f4f1621c62425c750211fd7321017d29cc1
changed the Create function signature.  What that commit does is
essentially: wrap the existing code with a backend interface to abstract
low-level reads and writes.  The raw file format is the only supported
backend and is what we're using.  Create() doesn't ask for the input
format because it now assumes the raw file backend (that we want).

This stmgr commit drops the format parameter Create() doesn't want
anymore.  If we run into other issues with Create() in the future, note
that the function may be deprecated in favor of doing your own create
based on file.CreateFromPath() followed by diskfs.OpenBackend().